SLIDER-1174 add API to simplify the use
