+++ app-packages/tensorflow/appConfig.default.json
@@ -5,8 +5,7 @@
   "global": {
     "site.global.hadoop.conf": "/etc/hadoop/conf",
     "site.global.yarn.cgroup.root": "/hadoop-yarn",
-    "site.global.user.scripts.entry": "mnist.py",
-    "site.global.user.data.dir": "hdfs://default/user/${USER_NAME}/mnist-data",
+    "site.global.user.scripts.entry": "ymnist.py",
     "site.global.user.checkpoint.prefix": "hdfs://default/user/${USER_NAME}/.slider/checkpoints",
     "site.global.user.name": "${USER_NAME}",
     "site.global.zookeeper.quorum": "${ZK_HOST}",
